backend/requirements.txt
fastapi>=0.111,<1.0
uvicorn>=0.30,<1.0
python-multipart>=0.0.9,<1.0
httpx>=0.27,<1.0
huggingface_hub>=0.31,<1.0
aiohttp>=3.11,<4.0
opencv-python-headless>=4.10,<5.0
numpy>=1.26,<3.0
python-dotenv>=1.0,<2.0
